Added exception handling + making other pipe types

+++ src/longpipes/PipeII.java
@@ -12,7 +12,7 @@ public class PipeII extends PipeI{
      * @param chemRes a boolean value representing if the pipe has chemical resistance 
 
      */    
-    public PipeII(int plasticGr,int len,int outerDi ,boolean chemRes){
+    public PipeII(int plasticGr,double len,double outerDi ,boolean chemRes){
         super(plasticGr,len,outerDi,chemRes);     
     }
